prov/efa: extract packet entry related items to a separate file
This patch move packet entry related structs and enums from rxr.h
to rxr_pkt_entry.h,

It also move paket entry related functions from rxr.h and rxr_ep.c
to rxr_pkt_entry.c.

It als rename functions to add a rxr_pkt_entry prefix. Such as:

    rxr_ep_get_pkt_entry -> rxr_pkt_entry_alloc
    rxr_copy_pkt_entry -> rxr_pkt_entry_copy
    rxr_release_tx_pkt_entry -> rxr_pkt_entry_release_tx
    rxr_release_rx_pkt_entry -> rxr_pkt_entry_release_rx
    rxr_ep_send_pkt -> rxr_pkt_entry_send
    rxr_ep_sendv_pkt -> rxr_pkt_entry_sendv
    rxr_ep_inject_pkt -> rxr_pkt_entry_inject
    rxr_ep_send_pkt_flags -> rxr_pkt_entry_send_with_flags
